refactor: WebP 파일로 이미지 포맷 변환
